Renaming path folder to get back in sync
