+++ client/components/category-single.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react'
 import { CategoryAllPaths, CategorySearch, CategoryPopularPaths } from './category'
 
 class CategorySinglePage extends Component {
-  
+
   render() {
     const categoryId = this.props.match.params.categoryId
     return (
