+++ client/components/paths/SinglePath.js
