+++ client/components/paths/index.js
