+++ client/components/paths/path-card-small.js
