+++ client/components/paths/path-card.js
